feat: Update passort.js
