+++ config/passport.js
@@ -1,6 +1,6 @@
 const FacebookStrategy = require('passport-facebook').Strategy
 const GoogleStrategy = require('passport-google-oauth2').Strategy
-const bcrypt = require('bcryptjs')
+const checkUser = require('../helpers/checkUser')
 const User = require('../models/users')
 
 module.exports = (passport) => {
@@ -13,36 +13,7 @@ module.exports = (passport) => {
         callbackURL: process.env.FACEBOOK_CALLBACK,
         profileFields: ['email', 'displayName']
       },
-      (accessToken, refreshToken, profile, done) => {
-        // find and create user
-        User.findOne({
-          email: profile._json.email
-        }).then((user) => {
-          // 如果 email 不存在就建立新的使用者
-          if (!user) {
-            const randomPassword = Math.random().toString(36).slice(-8)
-            bcrypt.genSalt(12, (_err, salt) =>
-              bcrypt.hash(randomPassword, salt, (_err, hash) => {
-                const newUser = User({
-                  name: profile._json.name,
-                  email: profile._json.email,
-                  password: hash
-                })
-                newUser
-                  .save()
-                  .then((user) => {
-                    return done(null, user)
-                  })
-                  .catch((err) => {
-                    console.log(err)
-                  })
-              })
-            )
-          } else {
-            return done(null, user)
-          }
-        })
-      }
+      checkUser
     )
   )
 
@@ -55,36 +26,7 @@ module.exports = (passport) => {
         callbackURL: process.env.GOOGL_CALLBACK,
         profileFields: ['email', 'displayName']
       },
-      (accessToken, refreshToken, profile, done) => {
-        // find and create user
-        User.findOne({
-          email: profile._json.email
-        }).then((user) => {
-          // 如果 email 不存在就建立新的使用者
-          if (!user) {
-            const randomPassword = Math.random().toString(36).slice(-8)
-            bcrypt.genSalt(12, (_err, salt) =>
-              bcrypt.hash(randomPassword, salt, (_err, hash) => {
-                const newUser = User({
-                  name: profile._json.name,
-                  email: profile._json.email,
-                  password: hash
-                })
-                newUser
-                  .save()
-                  .then((user) => {
-                    return done(null, user)
-                  })
-                  .catch((err) => {
-                    console.log(err)
-                  })
-              })
-            )
-          } else {
-            return done(null, user)
-          }
-        })
-      }
+      checkUser
     )
   )
